+++ package.json
@@ -53,7 +53,7 @@
 		"node": ">= 14.15.0"
 	},
 	"lint-staged": {
-		"*.js, *.jsx": [
+		"*.{js,jsx,ts,tsx,md,html}": [
 			"eslint",
 			"eslint --fix",
 			"prettier --write"
