+++ src/components/splash/OrganizationTimeline.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const OrganizationTimeline = () => {
+	return (
+		<ul>
+			<li>ORG 1</li>
+		</ul>
+	);
+};
+
+export default OrganizationTimeline;
