--- src/components/splash/OrganizationTimeline.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const OrganizationTimeline = () => {
-  return (
-    <div>OrganizationTimeline</div>
-  )
-}
-
-export default OrganizationTimeline
